requirements.txt
Django==3.0.7
python-ffmpeg-video-streaming==0.1.8
ffprobe-python==1.0.3
gunicorn==20.0.4
psycopg2-binary==2.8.5